+++ mreg_cli/commands/host_submodules/rr.py
@@ -42,7 +42,6 @@ from __future__ import annotations
 import argparse
 
 from mreg_api.models import (
-    MX,
     NAPTR,
     SSHFP,
     TXT,
@@ -56,7 +55,6 @@ from mreg_api.models import (
     Srv,
 )
 from mreg_api.models.fields import HostName
-from prompt_toolkit.output import Output
 
 from mreg_cli.commands.host import registry as command_registry
 from mreg_cli.exceptions import (
@@ -363,7 +361,7 @@ def naptr_add(args: argparse.Namespace) -> None:
     existing_naptr = NAPTR.get_by_query_unique(params)
     if existing_naptr:
         raise EntityAlreadyExists(f"{host} already has that NAPTR defined.")
-    NAPTR.create(params=params)
+    NAPTR.create(data=params)
     OutputManager().add_ok(f"Added NAPTR record to {host.name}.")