+++ mreg_cli/output/host.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from collections.abc import Sequence
-from typing import Literal
+from typing import Literal, NamedTuple
 
 from mreg_api import MregClient
 from mreg_api.endpoints import Endpoint
@@ -46,6 +46,7 @@ def output_host(
     :param names: If True, output the host names only.
     :param traverse_hostgroups: If True, traverse hostgroups and include them.
     """
+    _ = names
     from mreg_cli.output.group import output_hostgroups  # noqa: PLC0415
     from mreg_cli.output.policy import output_roles  # noqa: PLC0415
 
@@ -118,19 +119,42 @@ def output_hostlist(hostlist: HostList) -> None:
     if not hostlist.results:
         raise EntityNotFound("No hosts found.")
 
-    max_name = max_contact = 20
-    for i in hostlist.results:
-        max_name = max(max_name, len(str(i.name)))
-        max_contact = max(max_contact, max((len(c) for c in i.contact_emails), default=0))
+    max_name = max_contacts = max_ips = 20
 
-    def _format(name: str, contact: str, comment: str) -> None:
+    class HostOutput(NamedTuple):
+        name: str
+        contacts: str
+        comment: str
+        ips: str
+
+    hosts = [
+        HostOutput(
+            name=str(i.name),
+            contacts=", ".join(i.contact_emails),
+            comment=i.comment or "",
+            ips=", ".join(str(ip.ipaddress) for ip in i.ipaddresses),
+        )
+        for i in hostlist.results
+    ]
+    max_name = max(max_name, max(len(host.name) for host in hosts))
+    max_contacts = max(max_contacts, max(len(host.contacts) for host in hosts))
+    max_ips = max(max_ips, max(len(host.ips) for host in hosts))
+
+    def _format(name: str, contact: str, ips: str, comment: str) -> None:
         OutputManager().add_line(
-            "{0:<{1}} {2:<{3}} {4}".format(name, max_name, contact, max_contact, comment)
+            "{0:<{1}} {2:<{3}} {4:<{5}} {6}".format(
+                name, max_name, contact, max_contacts, ips, max_ips, comment
+            )
         )
 
-    _format("Name", "Contact", "Comment")
-    for i in hostlist.results:
-        _format(str(i.name), ", ".join(i.contact_emails), i.comment)
+    _format("Name", "Contact", "IP", "Comment")
+    for host in hosts:
+        _format(
+            name=host.name,
+            contact=host.contacts,
+            ips=host.ips,
+            comment=host.comment,
+        )
 
 
 def output_host_networks(
@@ -271,6 +295,7 @@ def output_host_roles(host: Host, padding: int = 14) -> None:
     :param host: Host whose roles to output.
     :param padding: Number of spaces for left-padding the output.
     """
+    _ = padding
     roles = host.roles
     manager = OutputManager()
     if not roles: